separated tests more and added comments regarding the necessary active beans
