+++ src/test/java/com/example/demo/WebAndSecurity/PublicEndpointsAndLockedPrivateEndpointsTest.java
@@ -1,0 +1,216 @@
+package com.example.demo.WebAndSecurity;
+
+import com.example.demo.Matcher.TestUtils;
+import com.example.demo.matcher.Matcher;
+import com.example.demo.matcher.models.*;
+import com.example.demo.matcher.services.OrderService;
+import com.example.demo.matcher.services.TradeService;
+import com.example.demo.security.service.UserService;
+import com.example.demo.security.userInfo.AppUser;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.doReturn;
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+
+// Requires real security config and filter and nothing else.
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class PublicEndpointsAndLockedPrivateEndpointsTest {
+
+    @Autowired
+    private WebApplicationContext webApplicationContext;
+
+    @Autowired
+    private UserService userService;
+
+    private MockMvc mvc;
+
+    @MockBean
+    private OrderService orderService;
+
+    @MockBean
+    private TradeService tradeService;
+
+    @MockBean
+    private Matcher matcher;
+
+    private AppUser testUser1;
+    private AppUser testUser2;
+
+
+    @BeforeEach
+    public void setup() throws Exception {
+        this.mvc = MockMvcBuilders.webAppContextSetup(this.webApplicationContext).apply(springSecurity()).build();
+
+        testUser1 = new AppUser(null, "testName1", "testUsername1", "testPassword1", new ArrayList<>());
+        testUser2 = new AppUser(null, "testName2", "testUsername2", "testPassword2", new ArrayList<>());
+
+        userService.saveUser(testUser1);
+        userService.saveUser(testUser2);
+    }
+
+    // ========= Test that public endpoints are accessible =============================================================
+
+    static List<OrderbookItem> testOrderbook1() {
+        return List.of(
+                TestUtils.makeOrderbookItem(1, 1),
+                TestUtils.makeOrderbookItem(2, 2)
+        );
+    }
+    @Test
+    public void itShouldAllowAccessToPublicBuyOrderbook() throws Exception {
+        doReturn(testOrderbook1()).when(orderService).getOrderbook(OrderAction.BUY);
+
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/public/orderbook/buy"))
+                .andReturn();
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.OK.value());
+        assertThat(result.getResponse().getContentAsString()).isEqualTo(TestUtils.asJsonString(testOrderbook1()));
+    }
+
+    @Test
+    public void itShouldAllowAccessToPublicSellOrderbook() throws Exception {
+        doReturn(testOrderbook1()).when(orderService).getOrderbook(OrderAction.SELL);
+
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/public/orderbook/sell"))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.OK.value());
+        assertThat(result.getResponse().getContentAsString()).isEqualTo(TestUtils.asJsonString(testOrderbook1()));
+    }
+
+    @Test
+    public void itShouldAllowAccessToPublicBuyOrderdepth() throws Exception {
+        doReturn(testOrderbook1()).when(orderService).getOrderDepth(OrderAction.BUY);
+
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/public/orderbook/depth/buy"))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.OK.value());
+        assertThat(result.getResponse().getContentAsString()).isEqualTo(TestUtils.asJsonString(testOrderbook1()));
+    }
+
+    @Test
+    public void itShouldAllowAccessToPublicSellOrderdepth() throws Exception {
+        doReturn(testOrderbook1()).when(orderService).getOrderDepth(OrderAction.SELL);
+
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/public/orderbook/depth/sell"))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.OK.value());
+        assertThat(result.getResponse().getContentAsString()).isEqualTo(TestUtils.asJsonString(testOrderbook1()));
+    }
+
+    @Test
+    public void itShouldAllowAccessToTradebook() throws Exception {
+        List<Trade> testTradebook = TestUtils.makeRandomTradebook();
+        doReturn(testTradebook).when(tradeService).getRecent();
+
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/public/tradebook"))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.OK.value());
+        assertThat(result.getResponse().getContentAsString()).isEqualTo(TestUtils.asJsonString(testTradebook));
+    }
+
+    // ========= Test that private endpoints are inaccessible without a good token =====================================
+
+    @Test
+    public void itShouldNotAllowAccessToPrivateBuyOrdersWithoutAnyJWT() throws Exception {
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/private/orderbook/buy/" + testUser1.getUsername()))
+                .andReturn();
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+
+    @Test
+    public void itShouldNotAllowAccessToPrivateSellOrdersWithoutAnyJWT() throws Exception {
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/private/orderbook/Sell/" + testUser1.getUsername()))
+                .andReturn();
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+
+    @Test
+    void itShouldNotAllowCreationOfOrdersWithoutAnyJWT() throws Exception {
+        Order newOrder = TestUtils.makeOrder(1 ,1, "b");
+
+        NewOrderParams newOrderParams = new NewOrderParams(
+                newOrder.getUsername(),
+                newOrder.getPrice().doubleValue(),
+                newOrder.getQuantity().doubleValue(),
+                "buy"
+        );
+
+        // API call
+        MvcResult result = mvc.perform(MockMvcRequestBuilders.post("/private/make/order")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(TestUtils.asJsonString(newOrderParams)))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+
+    // I may split the below three tests later in case we want to check for different responses depending on WHY the jwt is invalid
+    // (badly formatted, expired, etc)
+
+    @Test
+    public void itShouldNotAllowAccessToPrivateBuyOrdersWithoutValidJWT() throws Exception {
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/private/orderbook/buy/" + testUser1.getUsername())
+                                .header("Authorization", "Bearer badtoken"))
+                .andReturn();
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+
+    @Test
+    public void itShouldNotAllowAccessToPrivateSellOrdersWithoutValidJWT() throws Exception {
+        MvcResult result = mvc.perform(
+                        MockMvcRequestBuilders.get("/private/orderbook/sell/" + testUser1.getUsername())
+                                .header("Authorization", "Bearer badtoken"))
+                .andReturn();
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+
+    @Test
+    void itShouldNotAllowCreationOfOrdersWithoutValidJWT() throws Exception {
+        Order newOrder = TestUtils.makeOrder(1 ,1, "b");
+
+        NewOrderParams newOrderParams = new NewOrderParams(
+                newOrder.getUsername(),
+                newOrder.getPrice().doubleValue(),
+                newOrder.getQuantity().doubleValue(),
+                "buy"
+        );
+
+        // API call
+        MvcResult result = mvc.perform(MockMvcRequestBuilders.post("/private/make/order")
+                        .header("Authorization", "Bearer badtoken")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(TestUtils.asJsonString(newOrderParams)))
+                .andReturn();
+
+        assertThat(result.getResponse().getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
+    }
+}
